stop using klog

+++ pkg/checker/podstartup/podstartup.go
@@ -14,7 +14,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/klog/v2"
 
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
@@ -65,8 +64,8 @@ func (c *PodStartupChecker) Name() string {
 func (c *PodStartupChecker) Run(ctx context.Context) error {
 	// labels shared by all synthetic pods created by this checker
 	podLabels := map[string]string{
-		"app":                                 "cluster-health-monitor-podstartup-synthetic",
 		"cluster-health-monitor/checker-name": c.name,
+		"app":                                 "cluster-health-monitor-podstartup-synthetic",
 	}
 
 	// garbage collect any synthetic pods previously created by this checker
@@ -76,16 +75,10 @@ func (c *PodStartupChecker) Run(ctx context.Context) error {
 	if err != nil {
 		return fmt.Errorf("failed to list pods in namespace %s: %w", c.config.Namespace, err)
 	}
-	go func() {
-		defer func() {
-			if r := recover(); r != nil {
-				klog.Errorf("panic in garbageCollect: %s", err.Error())
-			}
-		}()
-		if err := c.garbageCollect(ctx, pods.Items); err != nil {
-			klog.Warningf("garbageCollect failed: %s", err.Error())
-		}
-	}()
+
+	if err := c.garbageCollect(ctx, pods.Items); err != nil {
+		fmt.Printf("garbageCollect failed: %s\n", err.Error())
+	}
 
 	// do not run the checker if the maximum number of synthetic pods has been reached
 	if len(pods.Items) >= c.config.MaxSyntheticPods {
@@ -114,24 +107,28 @@ func (c *PodStartupChecker) Run(ctx context.Context) error {
 	}
 
 	// Poll the durations to calculate the startup duration metric
-	imagePullDuration, err := c.pollImagePullDuration(ctx, synthPod.Name)
-	if err != nil {
-		return fmt.Errorf("failed to poll image pull duration for pod %s in namespace %s: %w", synthPod.Name, c.config.Namespace, err)
-	}
 	podCreationToContainerReadyDuration, err := c.pollPodCreationToContainerReadyDuration(ctx, synthPod.Name)
 	if err != nil {
 		return fmt.Errorf("failed to poll pod creation and container ready time for pod %s in namespace %s: %w", synthPod.Name, c.config.Namespace, err)
 	}
-	podStartupTime := podCreationToContainerReadyDuration - imagePullDuration
+	// This does not poll because if the container is ready, the event for pulling the image should already exist.
+	imagePullDuration, err := c.getImagePullDuration(ctx, synthPod.Name)
+	if err != nil {
+		return fmt.Errorf("failed to poll image pull duration for pod %s in namespace %s: %w", synthPod.Name, c.config.Namespace, err)
+	}
 
-	// TODO: Record startup duration as Prometheus metric
-	fmt.Println(podStartupTime)
+	podStartupDuration := podCreationToContainerReadyDuration - imagePullDuration
+	if podStartupDuration > 5*time.Second {
+		// TODO return unhealthy result
+		return fmt.Errorf("pod startup duration for pod %s in namespace %s is too long", synthPod.Name, c.config.Namespace)
+	}
 
 	err = c.k8sClientset.CoreV1().Pods(c.config.Namespace).Delete(ctx, synthPod.Name, metav1.DeleteOptions{})
 	if err != nil && !apierrors.IsNotFound(err) {
-		klog.Warningf("failed to delete synthetic pod %s in namespace %s: %s", synthPod.Name, c.config.Namespace, err.Error())
+		fmt.Printf("failed to delete synthetic pod %s in namespace %s: %s\n", synthPod.Name, c.config.Namespace, err.Error())
 	}
 
+	// TODO return healthy result
 	return nil
 }
 
@@ -153,11 +150,11 @@ func (c *PodStartupChecker) pollPodCreationToContainerReadyDuration(ctx context.
 	err := wait.PollUntilContextCancel(ctx, 5*time.Second, false, func(ctx context.Context) (bool, error) {
 		pod, err := c.k8sClientset.CoreV1().Pods(c.config.Namespace).Get(ctx, podName, metav1.GetOptions{})
 		if err != nil {
-			klog.Warningf("failed to get pod %s in namespace %s: %s", podName, c.config.Namespace, err.Error())
+			fmt.Printf("failed to get pod %s in namespace %s: %s", podName, c.config.Namespace, err.Error())
 			return false, nil
 		}
 		if len(pod.Status.ContainerStatuses) == 0 {
-			klog.Warningf("pod %s in namespace %s has no container statuses", podName, c.config.Namespace)
+			fmt.Printf("pod %s in namespace %s has no container statuses\n", podName, c.config.Namespace)
 			return false, nil
 		}
 		for _, status := range pod.Status.ContainerStatuses {
@@ -168,7 +165,7 @@ func (c *PodStartupChecker) pollPodCreationToContainerReadyDuration(ctx context.
 				return true, nil
 			}
 		}
-		klog.Warningf("no running containers for pod %s in namespace %s", podName, c.config.Namespace)
+		fmt.Printf("no running containers for pod %s in namespace %s\n", podName, c.config.Namespace)
 		return false, nil
 	})
 	if err != nil {
@@ -177,43 +174,28 @@ func (c *PodStartupChecker) pollPodCreationToContainerReadyDuration(ctx context.
 	return podCreationToContainerReadyDuration, nil
 }
 
-// Polls pod events until an event with reason "Pulled" is found or the context is no longer valid. Returns the image pull duration
-// including waiting time.
-func (c *PodStartupChecker) pollImagePullDuration(ctx context.Context, podName string) (time.Duration, error) {
-	var imagePullDuration time.Duration
-	err := wait.PollUntilContextCancel(ctx, 5*time.Second, false, func(ctx context.Context) (bool, error) {
-		events, err := c.k8sClientset.CoreV1().Events(c.config.Namespace).List(ctx, metav1.ListOptions{
-			FieldSelector: fmt.Sprintf("involvedObject.name=%s,reason=Pulled", podName),
-		})
-		if err != nil {
-			klog.Errorf("failed to list events for pod %s in namespace %s: %s", podName, c.config.Namespace, err.Error())
-			return false, nil
-		}
-
-		// events with reason=Pulled have messages expected to be in one of two formats:
-		// 1. "Successfully pulled image \"k8s.gcr.io/pause:3.2\" in 426ms (426ms including waiting). Image size: 299513 bytes."
-		// 2. "Container image \"k8s.gcr.io/pause:3.2\" already present on machine"
-		for _, event := range events.Items {
-			if strings.Contains(event.Message, "Successfully pulled image") {
-				imagePullDuration, err = c.extractImagePullDuration(event.Message)
-				if err != nil {
-					return false, fmt.Errorf("failed to extract image pull duration from event message: %w", err)
-				}
-				return true, nil
-			} else if strings.Contains(event.Message, "already present on machine") {
-				imagePullDuration = 0
-				return true, nil
-			} else {
-				klog.Warningf("Unexpected event message format for pod %s in namespace %s: %s", podName, c.config.Namespace, event.Message)
-				return false, nil
-			}
-		}
-		return false, nil
+// Returns the image pull duration including waiting time.
+func (c *PodStartupChecker) getImagePullDuration(ctx context.Context, podName string) (time.Duration, error) {
+	events, err := c.k8sClientset.CoreV1().Events(c.config.Namespace).List(ctx, metav1.ListOptions{
+		FieldSelector: fmt.Sprintf("involvedObject.name=%s,reason=Pulled", podName),
 	})
 	if err != nil {
-		return 0, err
+		return 0, fmt.Errorf("failed to list events for pod %s in namespace %s: %w", podName, c.config.Namespace, err)
 	}
-	return imagePullDuration, nil
+
+	// events with reason=Pulled have messages expected to be in one of two formats:
+	// 1. "Successfully pulled image \"k8s.gcr.io/pause:3.2\" in 426ms (426ms including waiting). Image size: 299513 bytes."
+	// 2. "Container image \"k8s.gcr.io/pause:3.2\" already present on machine"
+	for _, event := range events.Items {
+		if strings.Contains(event.Message, "Successfully pulled image") {
+			return c.extractImagePullDuration(event.Message)
+		} else if strings.Contains(event.Message, "already present on machine") {
+			return 0, nil
+		} else {
+			fmt.Printf("Unexpected event message format for pod %s in namespace %s: %s\n", podName, c.config.Namespace, event.Message)
+		}
+	}
+	return 0, fmt.Errorf("no image pull events found for pod %s in namespace %s", podName, c.config.Namespace)
 }
 
 // Extracts the image pull duration in ms from the event message. Message is expected to be in the format:
@@ -222,11 +204,11 @@ func (c *PodStartupChecker) extractImagePullDuration(message string) (time.Durat
 	re := regexp.MustCompile(`\((\d+)ms including waiting\)`)
 	matches := re.FindStringSubmatch(message)
 	if len(matches) != 2 {
-		return 0, fmt.Errorf("message in unexpected format: %s", message)
+		return 0, fmt.Errorf("failed to extract image pull duration from event message: message in unexpected format: %s", message)
 	}
 	imagePullDuration, err := strconv.Atoi(matches[1])
 	if err != nil {
-		return 0, fmt.Errorf("could not parse duration: %s", matches[1])
+		return 0, fmt.Errorf("failed to extract image pull duration from event message: could not parse duration: %s", matches[1])
 
 	}
 	return time.Duration(imagePullDuration) * time.Millisecond, nil
